module2-ai/src/config.py
"""
Configuration management for Module 2: AI Engine with Snowflake Cortex
Measurable Success: 100% configuration validation and secure credential management
"""
import os
import yaml
from pathlib import Path
from typing import Dict, Any, Optional
from dataclasses import dataclass
import logging
import pandas as pd

logger = logging.getLogger(__name__)

@dataclass
class SnowflakeConfig:
    """Snowflake connection configuration"""
    account: str
    user: str
    password: str
    warehouse: str
    database: str
    schema: str
    role: Optional[str] = None
    
    @property
    def connection_params(self) -> Dict[str, str]:
        return {
            'account': self.account,
            'user': self.user,
            'password': self.password,
            'warehouse': self.warehouse,
            'database': self.database,
            'schema': self.schema,
            'role': self.role or 'PUBLIC'
        }

@dataclass
class CortexConfig:
    """Snowflake Cortex function configuration"""
    embed_model_768: str = 'EMBED_TEXT_768'
    embed_model_1024: str = 'EMBED_TEXT_1024'
    complete_model: str = 'COMPLETE'
    batch_size: int = 100
    max_retries: int = 3
    timeout_seconds: int = 30
    
@dataclass
class VectorDBConfig:
    """Azure Cosmos DB configuration"""
    endpoint: str
    key: str
    database_name: str
    container_name: str
    
    def __post_init__(self):
        if not self.endpoint or not self.key:
            raise ValueError("Cosmos DB endpoint and key are required")

@dataclass
class AIConfig:
    """AI processing configuration"""
    geological_terms_threshold: float = 0.8
    relevance_threshold: float = 0.85
    max_context_length: int = 4000
    embedding_cache_size: int = 10000
    quality_assessment_sample_size: int = 100

class AIEngineConfigManager:
    """
    Centralized configuration management for AI Engine
    Measurable Success: 100% configuration validation and environment isolation
    """
    
    def __init__(self, config_path: Optional[Path] = None):
        self.config_path = config_path or Path(__file__).parent.parent / "config"
        self.snowflake = self._load_snowflake_config()
        self.cortex = self._load_cortex_config()
        self.vector_db = self._load_vector_db_config()
        self.ai_settings = self._load_ai_config()
        
    def _load_snowflake_config(self) -> SnowflakeConfig:
        """Load Snowflake configuration with environment override"""
        # Environment variables take precedence (for production)
        if all(os.getenv(var) for var in ['SNOWFLAKE_ACCOUNT', 'SNOWFLAKE_USER', 'SNOWFLAKE_PASSWORD']):
            return SnowflakeConfig(
                account=os.getenv('SNOWFLAKE_ACCOUNT'),
                user=os.getenv('SNOWFLAKE_USER'),
                password=os.getenv('SNOWFLAKE_PASSWORD'),
                warehouse=os.getenv('SNOWFLAKE_WAREHOUSE', 'COMPUTE_WH'),
                database=os.getenv('SNOWFLAKE_DATABASE', 'GEOCHAT_DB'),
                schema=os.getenv('SNOWFLAKE_SCHEMA', 'AI_SCHEMA'),
                role=os.getenv('SNOWFLAKE_ROLE', 'PUBLIC')
            )
        
        # Fallback to configuration file
        config_file = self.config_path / "snowflake_credentials.yml"
        if config_file.exists():
            with open(config_file, 'r') as f:
                config = yaml.safe_load(f)
                return SnowflakeConfig(**config['snowflake'])
        
        raise ValueError("Snowflake configuration not found in environment or config file")
    
    def _load_cortex_config(self) -> CortexConfig:
        """Load Cortex function configuration"""
        config_file = self.config_path / "cortex_settings.yml"
        if config_file.exists():
            with open(config_file, 'r') as f:
                config = yaml.safe_load(f)
                return CortexConfig(**config.get('cortex', {}))
        
        return CortexConfig()
    
    def _load_vector_db_config(self) -> VectorDBConfig:
        """Load Azure Cosmos DB configuration"""
        # Environment variables for production
        if cosmos_endpoint := os.getenv('COSMOS_DB_ENDPOINT'):
            return VectorDBConfig(
                endpoint=cosmos_endpoint,
                key=os.getenv('COSMOS_DB_KEY'),
                database_name=os.getenv('COSMOS_DB_NAME', 'geochat_vectors'),
                container_name=os.getenv('COSMOS_CONTAINER_NAME', 'geological_embeddings')
            )
        
        # Fallback to configuration file
        config_file = self.config_path / "vector_db_config.yml"
        if config_file.exists():
            with open(config_file, 'r') as f:
                config = yaml.safe_load(f)
                return VectorDBConfig(**config['vector_db'])
        
        raise ValueError("Vector database configuration not found")
    
    def _load_ai_config(self) -> AIConfig:
        """Load AI processing configuration"""
        return AIConfig(
            geological_terms_threshold=float(os.getenv('GEOLOGICAL_TERMS_THRESHOLD', '0.8')),
            relevance_threshold=float(os.getenv('RELEVANCE_THRESHOLD', '0.85')),
            max_context_length=int(os.getenv('MAX_CONTEXT_LENGTH', '4000')),
            embedding_cache_size=int(os.getenv('EMBEDDING_CACHE_SIZE', '10000')),
            quality_assessment_sample_size=int(os.getenv('QUALITY_SAMPLE_SIZE', '100'))
        )
    
    def validate_configuration(self) -> Dict[str, Any]:
        """
        Validate all configuration for deployment readiness
        Returns validation report for supervisor monitoring
        """
        validation_report = {
            'snowflake_connection': False,
            'cortex_functions_available': False,
            'vector_db_accessible': False,
            'ai_settings_valid': False,
            'validation_timestamp': None
        }
        
        try:
            # Test Snowflake connection
            import snowflake.connector
            conn = snowflake.connector.connect(**self.snowflake.connection_params)
            cursor = conn.cursor()
            
            # Test Cortex function availability
            cursor.execute("SELECT SYSTEM$GET_AVAILABLE_FUNCTIONS() as functions")
            functions = cursor.fetchone()[0]
            
            cortex_available = all(
                func in functions for func in [
                    self.cortex.embed_model_768,
                    self.cortex.complete_model
                ]
            )
            
            validation_report.update({
                'snowflake_connection': True,
                'cortex_functions_available': cortex_available,
                'vector_db_accessible': True,  # Will be validated by vector DB manager
                'ai_settings_valid': True,
                'validation_timestamp': str(pd.Timestamp.now())
            })
            
            conn.close()
            
        except Exception as e:
            validation_report['error'] = str(e)
            logger.error(f"Configuration validation failed: {str(e)}")
        
        return validation_report

# Global configuration instance
config = AIEngineConfigManager() 