module2-ai/requirements.txt
# Core dependencies for Module 2: AI Engine with Snowflake Cortex
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1
azure-cosmos==4.5.1
azure-identity==1.14.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
faiss-cpu==1.7.4
nltk==3.8.1
spacy==3.7.2
geopandas==0.13.2
shapely==2.0.1
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0
pyyaml==6.0.1
python-dotenv==1.0.0
prometheus-client==0.17.1
pytest==7.4.2
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Performance monitoring
psutil==5.9.6
memory-profiler==0.61.0
